Move core image into system base
I think it's over the top having 3 different images when for our use case we never really need different system-bases when making the distro.  We also have removed the container images since they don't make any sense for what we are doing. I expect there to be issues with doing this but they will be ironed out.